+++ apps/docs/src/components/feedbacks/Toast.stories.tsx
@@ -16,15 +16,12 @@ type Story = StoryObj<typeof meta>;
  */
 export const Default: Story = {
   args: {
-    style: {
-      position: "absolute",
-    },
-    direction: "center",
     icon: (
       <IconWrapper>
         <MdInfoOutline />
       </IconWrapper>
     ),
+    direction: "left",
     children: "Your trial has expired. Please purchase to continue.",
     tools: (
       <RoundButton
@@ -42,19 +39,14 @@ export const Default: Story = {
 
 export const Text: Story = {
   args: {
-    style: {
-      position: "absolute",
-    },
-    direction: "center",
+    direction: "left",
     children: "Your trial has expired. Please purchase to continue.",
   },
 };
 
 export const Icon: Story = {
   args: {
-    style: {
-      position: "absolute",
-    },
+    direction: "left",
     icon: (
       <IconWrapper
         style={{
@@ -64,16 +56,12 @@ export const Icon: Story = {
         <MdInfoOutline />
       </IconWrapper>
     ),
-    direction: "center",
     children: "Your trial has expired. Please purchase to continue.",
   },
 };
 
 export const Closable: Story = {
   args: {
-    style: {
-      position: "absolute",
-    },
     closable: true,
     icon: (
       <IconWrapper
@@ -84,7 +72,7 @@ export const Closable: Story = {
         <MdInfoOutline />
       </IconWrapper>
     ),
-    direction: "center",
+    direction: "left",
     children: "Your trial has expired. Please purchase to continue.",
   },
 };
@@ -93,9 +81,6 @@ export const Variant: Story = {
   args: {
     closable: true,
     variant: "primary",
-    style: {
-      position: "absolute",
-    },
     icon: (
       <IconWrapper
         style={{
@@ -105,7 +90,7 @@ export const Variant: Story = {
         <MdInfoOutline />
       </IconWrapper>
     ),
-    direction: "center",
+    direction: "left",
     children: "Your trial has expired. Please purchase to continue.",
   },
 };
@@ -113,9 +98,6 @@ export const Variant: Story = {
 export const Tools: Story = {
   args: {
     closable: true,
-    style: {
-      position: "absolute",
-    },
     tools: (
       <RoundButton
         outline
@@ -135,7 +117,7 @@ export const Tools: Story = {
         <MdInfoOutline />
       </IconWrapper>
     ),
-    direction: "center",
+    direction: "left",
     children: "Your trial has expired. Please purchase to continue.",
   },
 };